+++ test/DefifaHookRegressions.t.sol
@@ -115,16 +115,16 @@ contract DefifaHookRegressions is JBTest, TestBaseWorkflow {
         governor = new DefifaGovernor(jbController(), address(this));
         JBAddressRegistry _registry = new JBAddressRegistry();
         DefifaTokenUriResolver _tokenUriResolver = new DefifaTokenUriResolver(address(this));
-        deployer = new DefifaDeployer({deployer: address(this), initialOwner: address(this)});
-        deployer.setChainSpecificConstants({
-            newHookCodeOrigin: address(hook),
-            newTokenUriResolver: _tokenUriResolver,
-            newGovernor: governor,
-            newController: jbController(),
-            newRegistry: _registry,
-            newDefifaProjectId: _defifaProjectId,
-            newBaseProtocolProjectId: _protocolFeeProjectId,
-            newHookStore: new JB721TiersHookStore()
+        deployer = new DefifaDeployer({
+            initialOwner: address(this),
+            hookCodeOrigin: address(hook),
+            tokenUriResolver: _tokenUriResolver,
+            governor: governor,
+            controller: jbController(),
+            registry: _registry,
+            defifaProjectId: _defifaProjectId,
+            baseProtocolProjectId: _protocolFeeProjectId,
+            hookStore: new JB721TiersHookStore()
         });
 
         hook.transferOwnership(address(deployer));