+++ test/DefifaUSDC.t.sol
@@ -136,16 +136,16 @@ contract DefifaUSDCTest is JBTest, TestBaseWorkflow {
         hook =
             new DefifaHook(jbDirectory(), IERC20(_defifaProjectTokenAccount), IERC20(_protocolFeeProjectTokenAccount));
         governor = new DefifaGovernor(jbController(), address(this));
-        deployer = new DefifaDeployer({deployer: address(this), initialOwner: address(this)});
-        deployer.setChainSpecificConstants({
-            newHookCodeOrigin: address(hook),
-            newTokenUriResolver: new DefifaTokenUriResolver(address(this)),
-            newGovernor: governor,
-            newController: jbController(),
-            newRegistry: new JBAddressRegistry(),
-            newDefifaProjectId: _defifaProjectId,
-            newBaseProtocolProjectId: _protocolFeeProjectId,
-            newHookStore: new JB721TiersHookStore()
+        deployer = new DefifaDeployer({
+            initialOwner: address(this),
+            hookCodeOrigin: address(hook),
+            tokenUriResolver: new DefifaTokenUriResolver(address(this)),
+            governor: governor,
+            controller: jbController(),
+            registry: new JBAddressRegistry(),
+            defifaProjectId: _defifaProjectId,
+            baseProtocolProjectId: _protocolFeeProjectId,
+            hookStore: new JB721TiersHookStore()
         });
 
         uint8[] memory permissionIds = new uint8[](1);