+++ test/regression/OneTierZeroTimeoutLock.t.sol
@@ -98,16 +98,16 @@ contract OneTierZeroTimeoutLockTest is JBTest, TestBaseWorkflow {
 
         hook = new DefifaHook(jbDirectory(), IERC20(defifaToken), IERC20(protocolFeeToken));
         governor = new DefifaGovernor(jbController(), address(this));
-        deployer = new DefifaDeployer({deployer: address(this), initialOwner: address(this)});
-        deployer.setChainSpecificConstants({
-            newHookCodeOrigin: address(hook),
-            newTokenUriResolver: new DefifaTokenUriResolver(address(this)),
-            newGovernor: governor,
-            newController: jbController(),
-            newRegistry: new JBAddressRegistry(),
-            newDefifaProjectId: defifaProjectId,
-            newBaseProtocolProjectId: protocolFeeProjectId,
-            newHookStore: new JB721TiersHookStore()
+        deployer = new DefifaDeployer({
+            initialOwner: address(this),
+            hookCodeOrigin: address(hook),
+            tokenUriResolver: new DefifaTokenUriResolver(address(this)),
+            governor: governor,
+            controller: jbController(),
+            registry: new JBAddressRegistry(),
+            defifaProjectId: defifaProjectId,
+            baseProtocolProjectId: protocolFeeProjectId,
+            hookStore: new JB721TiersHookStore()
         });
 
         hook.transferOwnership(address(deployer));