+++ test/regression/PendingReserveQuorumGrief.t.sol
@@ -110,16 +110,16 @@ contract PendingReserveQuorumGriefTest is JBTest, TestBaseWorkflow {
 
         _hookImpl = new DefifaHook(jbDirectory(), IERC20(defifaToken), IERC20(nanaToken));
         _governorImpl = new DefifaGovernor(jbController(), address(this));
-        _deployer = new DefifaDeployer({deployer: address(this), initialOwner: address(this)});
-        _deployer.setChainSpecificConstants({
-            newHookCodeOrigin: address(_hookImpl),
-            newTokenUriResolver: new DefifaTokenUriResolver(address(this)),
-            newGovernor: _governorImpl,
-            newController: jbController(),
-            newRegistry: new JBAddressRegistry(),
-            newDefifaProjectId: _defifaProjectId,
-            newBaseProtocolProjectId: _protocolFeeProjectId,
-            newHookStore: new JB721TiersHookStore()
+        _deployer = new DefifaDeployer({
+            initialOwner: address(this),
+            hookCodeOrigin: address(_hookImpl),
+            tokenUriResolver: new DefifaTokenUriResolver(address(this)),
+            governor: _governorImpl,
+            controller: jbController(),
+            registry: new JBAddressRegistry(),
+            defifaProjectId: _defifaProjectId,
+            baseProtocolProjectId: _protocolFeeProjectId,
+            hookStore: new JB721TiersHookStore()
         });
 
         _hookImpl.transferOwnership(address(_deployer));